added Wallet and Transactions
Wallet knows about the owner and can track the unspent money. Transaction contains all necessary data, and tracks the inputs and outputs for said transaction.
TransactionInput and Output is needed to model the money that is exchanged and possibly returned as part of the transaction.

The Blockchain related data is also extracted to a separate class, and logic to the relevant services.